Add MOV to MP4 conversion

+++ src/lib/video-converter.ts
@@ -52,7 +52,7 @@ export async function convertToMp4(
   await ff.writeFile(inputName, await fetchFile(file));
   await ff.exec(['-i', inputName, '-c:v', 'libx264', '-preset', 'fast', '-crf', '28', '-c:a', 'aac', '-movflags', '+faststart', outputName]);
 
-  const data = await ff.readFile(outputName);
+  const data = (await ff.readFile(outputName)) as Uint8Array;
   // Cleanup
   await ff.deleteFile(inputName);
   await ff.deleteFile(outputName);
